Move FritzBoxBaseSwitch to switch.py
