+++ homeassistant/components/fritz/common.py
@@ -28,7 +28,7 @@ from homeassistant.helpers.device_registry import CONNECTION_NETWORK_MAC
 from homeassistant.helpers.dispatcher import dispatcher_send
 from homeassistant.helpers.entity import DeviceInfo
 from homeassistant.helpers.event import async_track_time_interval
-from homeassistant.util import dt as dt_util, slugify
+from homeassistant.util import dt as dt_util
 
 from .const import (
     DEFAULT_HOST,
@@ -352,84 +352,3 @@ class FritzBoxBaseEntity:
             "model": self._fritzbox_tools.model,
             "sw_version": self._fritzbox_tools.sw_version,
         }
-
-
-class FritzBoxBaseSwitch:
-    """Fritz switch base class."""
-
-    def __init__(self, fritzbox_tools: FritzBoxTools, switch_info: SwitchInfo) -> None:
-        """Init Fritzbox port switch."""
-        self.fritzbox_tools: FritzBoxTools = fritzbox_tools
-
-        self._description = switch_info["description"]
-        self._friendly_name = switch_info["friendly_name"]
-        self._icon = switch_info["icon"]
-        self._type = switch_info["type"]
-        self._update: Callable = switch_info["callback_update"]
-        self._switch: Callable = switch_info["callback_switch"]
-
-        self._name = f"{self._friendly_name} {self._description}"
-        self._unique_id = (
-            f"{self.fritzbox_tools.unique_id}-{slugify(self._description)}"
-        )
-
-        self._attributes: dict[str, str] = {}
-        self._is_available = True
-
-        self._attr_is_on = False
-
-    @property
-    def name(self):
-        """Return name."""
-        return self._name
-
-    @property
-    def icon(self):
-        """Return name."""
-        return self._icon
-
-    @property
-    def unique_id(self):
-        """Return unique id."""
-        return self._unique_id
-
-    @property
-    def device_info(self):
-        """Return the device information."""
-
-        return {
-            "connections": {(CONNECTION_NETWORK_MAC, self.fritzbox_tools.mac)},
-            "name": self._friendly_name,
-            "manufacturer": "AVM",
-            "model": self.fritzbox_tools.model,
-            "sw_version": self.fritzbox_tools.sw_version,
-        }
-
-    @property
-    def available(self) -> bool:
-        """Return availability."""
-        return self._is_available
-
-    @property
-    def extra_state_attributes(self):
-        """Return device attributes."""
-        return self._attributes
-
-    async def async_update(self):
-        """Update data."""
-        _LOGGER.debug("Updating '%s' (%s) switch state", self.name, self._type)
-        await self._update()
-
-    async def async_turn_on(self, **kwargs) -> None:
-        """Turn on switch."""
-        await self._async_handle_turn_on_off(turn_on=True)
-
-    async def async_turn_off(self, **kwargs) -> None:
-        """Turn off switch."""
-        await self._async_handle_turn_on_off(turn_on=False)
-
-    async def _async_handle_turn_on_off(self, turn_on: bool) -> bool:
-        """Handle switch state change request."""
-        await self._switch(turn_on)
-        self._attr_is_on = turn_on
-        return True
